+++ js/test/name/test/testname.js
@@ -221,7 +221,7 @@ function testNameZHPrefix() {
 }
 
 function testNameZHMarriedName() {
-    var name = new Name("王杨凭平", {locale: "zh-CN"});
+    var name = new Name("王杨凭平", {locale: "zh-CN", compoundFamilyName: true});
     assertNotUndefined(name);
     
     assertObjectContains({ givenName: "凭平", familyName: "王杨" }, name);